Content

+++ src/components/feed/Comment.js
@@ -4,10 +4,20 @@ import { Link } from "react-router-dom"
 import styled from "styled-components"
 import { FatText } from "../shared"
 import { useMutation, gql } from "@apollo/client"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+import { faPencilAlt, faTrash } from "@fortawesome/free-solid-svg-icons"
 
 const DELETE_COMMENT_MUTATION = gql`
 	mutation deleteComment($id: Int!) {
 		deleteComment(id: $id) {
+			ok
+		}
+	}
+`
+
+const EDIT_COMMENT_MUTATION = gql`
+	mutation editComment($id: Int!, $payload: String!) {
+		editComment(id: $id, payload: $payload) {
 			ok
 		}
 	}
@@ -19,6 +29,7 @@ const CommentContainer = styled.div`
 
 const CommentCaption = styled.span`
 	margin-left: 10px;
+	color: red;
 	a {
 		background-color: inherit;
 		color: ${(props) => props.theme.accent};
@@ -29,11 +40,18 @@ const CommentCaption = styled.span`
 	}
 `
 
+const Icons = styled.div``
+
+const Icon = styled.span`
+	cursor: pointer;
+	margin-left: 10px;
+`
+
 function Comment({ id, photoId, isMine, author, payload }) {
 	const updateDeleteComment = (cache, result) => {
 		const {
 			data: {
-				deleteComment: { ok },
+				deleteComment: { ok, error },
 			},
 		} = result
 		if (ok) {
@@ -48,14 +66,38 @@ function Comment({ id, photoId, isMine, author, payload }) {
 			})
 		}
 	}
+	const updateEditComment = (cache, result) => {
+		const {
+			data: {
+				editComment: { ok },
+			},
+		} = result
+		if (ok) {
+			console.log("ok")
+		}
+	}
 	const [deleteCommentMutation] = useMutation(DELETE_COMMENT_MUTATION, {
 		variables: {
 			id,
 		},
 		update: updateDeleteComment,
 	})
-	const onDeleteClick = () => {
-		deleteCommentMutation()
+	const [editCommentMutation] = useMutation(EDIT_COMMENT_MUTATION, {
+		variables: {
+			id,
+			payload,
+		},
+		update: updateEditComment,
+	})
+
+	const onDeleteClick = async () => {
+		const ok = window.confirm("Are you sure you want to delete this comment?")
+		if (ok) {
+			deleteCommentMutation()
+		}
+	}
+	const onEditClick = () => {
+		editCommentMutation()
 	}
 	return (
 		<CommentContainer>
@@ -75,7 +117,16 @@ function Comment({ id, photoId, isMine, author, payload }) {
 					)
 				)}
 			</CommentCaption>
-			{isMine ? <button onClick={onDeleteClick}>❌</button> : null}
+			{isMine ? (
+				<Icons>
+					<Icon onClick={onEditClick}>
+						<FontAwesomeIcon icon={faPencilAlt} />
+					</Icon>
+					<Icon onClick={onDeleteClick}>
+						<FontAwesomeIcon icon={faTrash} />
+					</Icon>
+				</Icons>
+			) : null}
 		</CommentContainer>
 	)
 }
